+++ sandbox/rick/nolong/kernel/platform/unix/FileSystem.cpp
@@ -721,7 +721,7 @@ BOOL SysFileIsPipe(STREAM_INFO * stream_info)
 }
 
 
-LONG SysTellPosition(STREAM_INFO * stream_info)
+size_t SysTellPosition(STREAM_INFO * stream_info)
 {                                                 /* flags.std has been changed */
    if (stream_info->flags.bstd || stream_info->fh == stdin_handle) return -1;
     return ftell(stream_info->stream_file);
